// line-ux/src/settings/title/title.types.ts
import * as echarts from 'echarts';

export type TitleType = echarts.EChartTitleOption;

export enum TitleEViewKey {
  Show = 'titleShow',
  //
  Link = 'titleLink',
  Target = 'titleTarget',
  //
  Html = 'titleHtml',
  Text = 'titleText',
  FontSize = 'titleFontSize',
  Color = 'titleColor',
  Bold = 'titleBold',
  Italic = 'titleItalic',
  LineHeight = 'titleLineHeight',
  TextAlign = 'titleTextAlign',
  Rich = 'titleRich',
  //
  Left = 'titleLeft',
  Top = 'titleTop',
  Right = 'titleRight',
  Bottom = 'titleBottom',
}